# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LLM Data Vault API
  description: |
    Enterprise-grade secure data management API for LLM training pipelines.

    ## Authentication

    Most endpoints require authentication via JWT bearer token or API key.

    ### JWT Authentication
    1. Call `/api/v1/auth/login` with credentials
    2. Use the returned `access_token` in the `Authorization` header
    3. Refresh tokens using `/api/v1/auth/refresh` before expiration

    ### API Key Authentication
    Include your API key in the `X-API-Key` header.

    ## Rate Limiting

    - Default: 100 requests/second per client
    - Burst: 200 requests
    - Rate limit headers included in responses

    ## Pagination

    List endpoints support cursor-based pagination:
    - `limit`: Number of items (default: 20, max: 100)
    - `cursor`: Cursor from previous response

    ## Error Responses

    All errors return JSON with:
    - `code`: Machine-readable error code
    - `message`: Human-readable description
    - `request_id`: For support inquiries

  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
    url: https://docs.example.com
  license:
    name: LLMDevOps-PSACL v1.0
    url: https://example.com/license
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging-api.example.com
    description: Staging
  - url: http://localhost:8080
    description: Local Development
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/webhooks/{webhookId}:
    parameters:
      - $ref: '#/components/parameters/WebhookId'

    get:
      tags: [Webhooks]
      summary: Get webhook
      description: Returns webhook details.
      operationId: getWebhook
      responses:
        '200':
          description: Webhook details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Webhooks]
      summary: Update webhook
      description: Updates webhook configuration.
      operationId: updateWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookRequest'
      responses:
        '200':
          description: Webhook updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Webhooks]
      summary: Delete webhook
      description: Deletes a webhook subscription.
      operationId: deleteWebhook
      responses:
        '204':
          description: Webhook deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  parameters:
...
    WebhookId:
      name: webhookId
      in: path
      required: true
      description: Webhook unique identifier
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Invalid or expired token
            request_id: req_abc123
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: Dataset with id 'ds_123' not found
            request_id: req_abc123
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: Validation failed
            errors:
              email:
                - must be a valid email address
              password:
                - must be at least 8 characters
            request_id: req_abc123
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        request_id:
          type: string
          description: Request ID for support
        timestamp:
          type: string
          format: date-time
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
...
    UpdateWebhookRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        secret:
          type: string
    WebhookResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...